Make Document::postTask() safe to call from a background thread https://bugs.webkit.org/show_bug.cgi?id=199585
Reviewed by Alex Christensen.

Make Document::postTask() safe to call from a background thread by not calling makeWeakPtr() on the Document.
Calling makeWeakPtr() on a document from a background thread is not safe since Document is a main thread
object. Instead, capture Document::identifier() in the lambda and lookup the document from its identifier
once we're on the main thread.

* dom/Document.cpp:
(WebCore::Document::postTask):

Identifier: 213489@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247239 268f45cc-cd09-0410-ab3c-d52691b4dbfc