[GTK] No web view updates after swapping web process if AC mode is forced https://bugs.webkit.org/show_bug.cgi?id=199826
Reviewed by Michael Catanzaro.

This is because when AC mode is forced the provisional page creates a drawing area that is always in AC
mode. When swapping processes, the current drawing area is destroyed (which exists AC mode) and the provisional
one is set. From that point the web process is in AC mode but the UI process isn't. We need to update the
compositing mode when swapping processes.

* UIProcess/API/gtk/WebKitWebViewBase.cpp:
(webkitWebViewBaseDidExitWebProcess):
* UIProcess/CoordinatedGraphics/DrawingAreaProxyCoordinatedGraphics.h:

Identifier: 213728@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@247506 268f45cc-cd09-0410-ab3c-d52691b4dbfc